requirements.txt
aiohttp==3.11.11
python-telegram-bot==22.0
solana==0.36.6
solders==0.26.0
Telethon==1.37.0